Project/Code refactor for more structural clarity
