+++ src/PizzaBar/CompareById.java
@@ -2,4 +2,5 @@ package PizzaBar;
 
 public class CompareById {
     //TODO: make a compare() method to sort orders by id
+    // TODO: !NEW! Maybe implement this in another helper/logic class?
 }
