+++ src/PizzaBar/CompareByTime.java
@@ -4,4 +4,5 @@ import java.util.Comparator;
 public class CompareByTime {
     //TODO: implement Comparator<Order>
     //TODO: make a compare() method to sort orders by pickupTime.
+    // TODO: !NEW! Maybe implement this in another helper/logic class?
 }
