--- src/PizzaBar/OrderHelper.java
@@ -1,5 +1,0 @@
-package PizzaBar;
-
-public class OrderHelper {
-    //TODO: import scanner, and make a scanner for user input
-}
