+++ src/PizzaBar/app/App.java
@@ -1,7 +1,8 @@
-package PizzaBar;
+package PizzaBar.app;
 
 public class App {
     public static void main(String[] args) {
         //TODO: call methods to run app
+        new Application().run();
     }
 }
