+++ src/PizzaBar/helpers/OrderHelper.java
@@ -1,0 +1,11 @@
+package PizzaBar.helpers;
+
+import PizzaBar.products.*;
+import PizzaBar.logic.*;
+
+import java.util.Scanner;
+import java.time.LocalDateTime;
+
+public class OrderHelper {
+    //TODO: import scanner, and make a scanner for user input
+}
