+++ src/PizzaBar/logic/OrderManager.java
@@ -1,4 +1,6 @@
-package PizzaBar;
+package PizzaBar.logic;
+
+import PizzaBar.products.*;
 
 public class OrderManager {
