+++ src/PizzaBar/products/OrderStatus.java
@@ -1,4 +1,4 @@
-package PizzaBar;
+package PizzaBar.products;
 
 public enum OrderStatus {
     //TODO: order statuses
