+++ src/PizzaBar/products/PizzaType.java
@@ -1,0 +1,4 @@
+package PizzaBar.products;
+
+public enum PizzaType {
+}
